requirements.txt
python-dotenv
flask
flask-cors
IP2Location
uwsgi